requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
py-clob-client==0.23.0
google-generativeai==0.3.2
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4
matplotlib==3.8.2
plotly==5.17.0
websockets==12.0
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
asyncio-mqtt==0.16.1
schedule==1.2.0
psutil==5.9.6
rich==13.7.0
typer==0.9.0
pydantic==2.5.0
pydantic-settings==2.1.0